+++ core/tauri/src/protocol/mod.rs
@@ -4,4 +4,6 @@
 
 #[cfg(feature = "protocol-asset")]
 pub mod asset;
+#[cfg(feature = "isolation")]
+pub mod isolation;
 pub mod tauri;
